fix: clean up the files and extract types and interfaces
